sdm660: use lazytime for /system and /vendor
lazytime reduces writes to disk by maintaining changes to inode
timestamps (access, modification and creation times) only in memory. The
on-disk timestamps are updated only when either (1) the file inode needs
to be updated for some change unrelated to file timestamps, (2) a sync
to disk occurs, (3) an undeleted inode is evicted from memory or (4) if
more than 24 hours passed since the the last time the in-memory copy was
written to disk.

Source:
https://wiki.archlinux.org/index.php/fstab#Kernel_name_descriptors

Signed-off-by: Panchajanya1999 <rsk52959@gmail.com>